Fixed potential streamer crash by fixing the non-determinism issues in the Nanite builder. Because of assumptions in the DDC system about cooking being deterministic, we could under various circumstances end up in a situations where StaticMesh and the associated Nanite streaming data was coming from two different cooks. This could easily cause a crash as the StaticMesh chunk contains explicit page and offset references into the streaming data.
Determinism fixes:
-Added sorting step during initial clustering in the case where an edge is shared by more than 2 triangles.
-Added sorting step to sort parent clusters after they have been generated.
-Changed cluster GUID from 32bit to 64bit. When cooking CitySample there was one case where two clusters that were compared ended up with the same GUID, resulting in undefined order.
-Initialized GUID in the cluster constructor that takes a list of clusters to merge. Before it would just be 0.
-Added missing initialization of vertex color on fallback meshes that don't use vertex color.

#jira UE-146347
#rb brian.karis, michal.valient
#preflight 6238817789625f06129f616f

[CL 19458684 by Rune Stubbe in 5.0 branch]

+++ Engine/Source/Developer/NaniteBuilder/Private/Cluster.h
@@ -190,7 +190,7 @@ public:
 	TMap< uint32, uint32 >	AdjacentClusters;
 
 	FBounds		Bounds;
-	uint32		GUID = 0;
+	uint64		GUID = 0;
 	int32		MipLevel = 0;
 
 	FIntVector	QuantizedPosStart		= { 0u, 0u, 0u };
